style: :lipstick: CSS has been implemented.
I forgot to upload a commit after finishing the HTML. I began working on the CSS right away, but when I remembered, I left the CSS file empty to quickly upload the HTML changes. Later, I had to cut the CSS changes in half to balance the versions. Unfortunately, it took me an hour to fix the exercise menu using flexbox in the CSS file.